feat: Modal props type 분리
