+++ components/src/lib/type/index.ts
@@ -1,0 +1,6 @@
+export interface ModalProps {
+  modalPosition: "center" | "bottom";
+  closeButtonPosition: "top" | "bottom";
+  placeholder?: string;
+  onSubmit?: (value: string) => void | string;
+}
